pyproject.toml
[project]
name = "data-master-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "fuzzywuzzy>=0.18.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "python-levenshtein>=0.27.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "xlrd>=2.0.2",
]
